requirements.txt
django-bootstrap3
factory_boy
django-extensions
django-bootstrap-datepicker-plus
freezegun